requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.43
Pillow==11.3.0
pytesseract==0.3.13
reportlab==4.4.3
python-dotenv==1.1.1
Jinja2==3.1.6
Werkzeug==3.1.3
click==8.1.8
itsdangerous==2.2.0
blinker==1.9.0
